code/requirements.txt
discord.py
discord-py-slash-command
python-dotenv
pytest